Make getCurrentChainConfig a static method of Etherscan
